fastapi-app/requirements.txt
fastapi>=0.95.2,<0.96.0
uvicorn>=0.22.0,<0.23.0
pydantic>=1.10.7,<1.11.0
httpx>=0.24.1,<0.25
pytest>=7.3.0,<7.4.0
pytest-asyncio>=0.21.0,<0.22.0
PyJWT>=2.7.0,<2.8.0